+++ core/transaction/sell_all_coin.go
@@ -45,7 +45,7 @@ func (data SellAllCoinData) TotalSpend(tx *Transaction, context *state.State) (T
 			}
 		}
 
-		if err := CheckForCoinSupplyOverflow(coin.Volume(), value); err != nil {
+		if err := CheckForCoinSupplyOverflow(coin.Volume(), value, coin.MaxSupply()); err != nil {
 			return nil, nil, nil, &Response{
 				Code: code.CoinSupplyOverflow,
 				Log:  err.Error(),
@@ -111,7 +111,7 @@ func (data SellAllCoinData) TotalSpend(tx *Transaction, context *state.State) (T
 			}
 		}
 
-		if err := CheckForCoinSupplyOverflow(coinTo.Volume(), value); err != nil {
+		if err := CheckForCoinSupplyOverflow(coinTo.Volume(), value, coinTo.MaxSupply()); err != nil {
 			return nil, nil, nil, &Response{
 				Code: code.CoinSupplyOverflow,
 				Log:  err.Error(),